requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.0